+++ src/waka-orchestrator/adaptors/keyvalueLocal.ts
@@ -1,7 +1,7 @@
 import * as path from 'path'
 import { promises as fs } from 'fs'
 import logger from '../logger'
-import BaseKeyvalue from './BaseKeyvalue'
+import { BaseKeyvalue } from '../../typings'
 
 // this is designed to be slow af to emulate the dynamoDB lag
 const filePath = path.join(__dirname, '../../cache/keyvalue-')
@@ -15,7 +15,6 @@ class KeyvalueLocal extends BaseKeyvalue {
   async read() {
     const { name } = this
     try {
-      console.log(filePath, name)
       const content = await fs.readFile(`${filePath}${name}.json`)
       const jsonstring = content.toString()
       const data = JSON.parse(jsonstring)
